Calibration is now possible without a calibration target.
Instead a dumbbell (object with two points of known respective distance) can be moved around the focus domain and recorded.

The algorithm is VERY slow, but it is robust (blow-up is hardly possible anymore) and eventually converges.

The file dumbbell.par controls the algorithm. Explanations will follow shortly on the ptv.origo in the manual pages.

git-svn-id: https://svn.origo.ethz.ch/ptv/trunk@75 dcddef73-2b9d-49fe-a651-d05a6fa2f102